pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "chemformer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.1",
    "httpx>=0.28.1",
    "joblib>=1.5.3",
    "liger-kernel>=0.6.4",
    "lmdb>=1.7.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-benchmark>=5.2.3",
    "pyyaml>=6.0.3",
    "rdkit>=2025.9.3",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "ruff>=0.5.5",
    "poethepoet>=0.27.0",
    "hydra-core>=1.2.0",
    "maturin>=1.10.2",
]

[tool.maturin]
manifest-path = "chemformer_rs/Cargo.toml"
features = ["pyo3/extension-module"]

[tool.ruff]
line-length = 120

[tool.ruff.format]
indent-style = "tab"

[tool.poe.tasks.build-rust]
cmd = "maturin develop --release --manifest-path chemformer_rs/Cargo.toml"
help = "Build the rust extension in release mode."

[tool.pytest.ini_options]
addopts = ["-m", "not benchmark"]
markers = ["benchmark: marks tests as benchmarks (deselect with '-m \"not benchmark\"')"]
